code cleanup: remove unecessary casts, duplicate semicolons, add @Override, use for loops etc.
git-svn-id: https://svn.codehaus.org/mule/branches/mule-3.x@22391 bf997673-6b11-0410-b953-e057580c5b09

+++ transports/tcp/src/main/java/org/mule/transport/tcp/PollingTcpMessageReceiver.java
@@ -70,7 +70,7 @@ public class PollingTcpMessageReceiver extends AbstractPollingMessageReceiver
         Socket socket = connector.getSocket(endpoint);
         try
         {
-            Object result = TcpMessageDispatcher.receiveFromSocket(socket, (int) timeout, endpoint);
+            Object result = TcpMessageDispatcher.receiveFromSocket(socket, timeout, endpoint);
             if (!(result == null))
             {
                 this.routeMessage(new DefaultMuleMessage(result, connector.getMuleContext()));
